+++ Tests/DatadogTests/Datadog/DatadogConfigurationBuilderTests.swift
@@ -66,6 +66,8 @@ class DatadogConfigurationBuilderTests: XCTestCase {
             XCTAssertEqual(configuration.additionalConfiguration.count, 0)
             XCTAssertNil(configuration.encryption)
             XCTAssertNil(configuration.serverDateProvider)
+            XCTAssertEqual(configuration.tracingHeaderTypes.count, 1)
+            XCTAssertEqual(configuration.tracingHeaderTypes.first, .dd)
         }
     }
 
@@ -185,6 +187,8 @@ class DatadogConfigurationBuilderTests: XCTestCase {
             XCTAssertEqual(configuration.proxyConfiguration?[kCFProxyPasswordKey] as? String, "proxypass")
             XCTAssertTrue(configuration.encryption is DataEncryptionMock)
             XCTAssertTrue(configuration.serverDateProvider is ServerDateProviderMock)
+            XCTAssertEqual(configuration.tracingHeaderTypes.count, 1)
+            XCTAssertEqual(configuration.tracingHeaderTypes.first, .dd)
         }
 
         XCTAssertTrue(rumConfigurationWithDefaultValues.rumUIKitViewsPredicate is DefaultUIKitRUMViewsPredicate)